moved kmalloc code to kernel/mm.

+++ libc++/dynmem.cc
@@ -1,13 +1,13 @@
-#include <stdlib.h>
+#include <cstdlib>
 
 // overload the operator "new"
-void *operator new(unsigned int size)
+void *operator new(size_t size)
 {
     return kmalloc(size);
 }
 
 // overload the operator "new[]"
-void *operator new[](unsigned int size)
+void *operator new[](size_t size)
 {
     return kmalloc(size);
 }
@@ -19,7 +19,7 @@ void operator delete(void *p)
 }
 
 //overload the operator "delete[]"
-void operator delete[](void * p)
+void operator delete[](void *p)
 {
     kfree(p);
 }
